Assignment 3

+++ routes/dishRouter.js
@@ -19,8 +19,9 @@ dishRouter.route('/')
             }, (err) => next(err))
             .catch((err) => next(err));
     })
-    .post(authenticate.verifyUser, (req, res, next) => {
-        Dishes.create(req.body)
+    .post(authenticate.verifyUser, authenticate.verifyAdmin, (req, res, next) => {
+        if(authenticate.verifyAdmin)
+           Dishes.create(req.body)
             .then((dish) => {
                 console.log('Dish Created ', dish);
                 res.statusCode = 200;
@@ -29,11 +30,12 @@ dishRouter.route('/')
             }, (err) => next(err))
             .catch((err) => next(err));
     })
-    .put(authenticate.verifyUser, (req, res, next) => {
+    .put(authenticate.verifyUser, authenticate.verifyAdmin, (req, res, next) => {
         res.statusCode = 403;
         res.end('PUT operation not supported on /dishes');
     })
-    .delete(authenticate.verifyUser, (req, res, next) => {
+    .delete(authenticate.verifyUser, authenticate.verifyAdmin, (req, res, next) => {
+        if (authenticate.verifyAdmin)
         Dishes.remove({})
             .then((resp) => {
                 res.statusCode = 200;
@@ -44,9 +46,11 @@ dishRouter.route('/')
     });
 
 
+
 dishRouter.route('/:dishId')
     .get((req, res, next) => {
         Dishes.findById(req.params.dishId)
+            .populate('comments.author')
             .then((dish) => {
                 res.statusCode = 200;
                 res.setHeader('Content-Type', 'application/json');
@@ -54,11 +58,12 @@ dishRouter.route('/:dishId')
             }, (err) => next(err))
             .catch((err) => next(err));
     })
-    .post(authenticate.verifyUser, (req, res, next) => {
+    .post(authenticate.verifyUser, authenticate.verifyAdmin, (req, res, next) => {
         res.statusCode = 403;
         res.end('POST operation not supported on /dishes/' + req.params.dishId);
     })
-    .put(authenticate.verifyUser, (req, res, next) => {
+    .put(authenticate.verifyUser, authenticate.verifyAdmin, (req, res, next) => {
+        if (authenticate.verifyAdmin)
         Dishes.findByIdAndUpdate(req.params.dishId, {
             $set: req.body
         }, { new: true })
@@ -69,7 +74,8 @@ dishRouter.route('/:dishId')
             }, (err) => next(err))
             .catch((err) => next(err));
     })
-    .delete(authenticate.verifyUser, (req, res, next) => {
+    .delete(authenticate.verifyUser, authenticate.verifyAdmin, (req, res, next) => {
+        if (authenticate.verifyAdmin)
         Dishes.findByIdAndRemove(req.params.dishId)
             .then((resp) => {
                 res.statusCode = 200;
@@ -82,6 +88,7 @@ dishRouter.route('/:dishId')
 dishRouter.route('/:dishId/comments')
     .get((req, res, next) => {
         Dishes.findById(req.params.dishId)
+            .populate('comments.author')
             .then((dish) => {
                 if (dish != null) {
                     res.statusCode = 200;
@@ -100,12 +107,17 @@ dishRouter.route('/:dishId/comments')
         Dishes.findById(req.params.dishId)
             .then((dish) => {
                 if (dish != null) {
+                    req.body.author = req.user._id;
                     dish.comments.push(req.body);
                     dish.save()
                         .then((dish) => {
-                            res.statusCode = 200;
-                            res.setHeader('Content-Type', 'application/json');
-                            res.json(dish);
+                            Dishes.findById(dish._id)
+                                .populate('comments.author')
+                                .then((dish) => {
+                                    res.statusCode = 200;
+                                    res.setHeader('Content-Type', 'application/json');
+                                    res.json(dish);
+                                })
                         }, (err) => next(err));
                 }
                 else {
@@ -116,6 +128,7 @@ dishRouter.route('/:dishId/comments')
             }, (err) => next(err))
             .catch((err) => next(err));
     })
+
     .put(authenticate.verifyUser, (req, res, next) => {
         res.statusCode = 403;
         res.end('PUT operation not supported on /dishes/'
@@ -147,6 +160,7 @@ dishRouter.route('/:dishId/comments')
 dishRouter.route('/:dishId/comments/:commentId')
     .get((req, res, next) => {
         Dishes.findById(req.params.dishId)
+            .populate('comments.author')
             .then((dish) => {
                 if (dish != null && dish.comments.id(req.params.commentId) != null) {
                     res.statusCode = 200;
@@ -166,15 +180,20 @@ dishRouter.route('/:dishId/comments/:commentId')
             }, (err) => next(err))
             .catch((err) => next(err));
     })
+
     .post(authenticate.verifyUser, (req, res, next) => {
         res.statusCode = 403;
         res.end('POST operation not supported on /dishes/' + req.params.dishId
             + '/comments/' + req.params.commentId);
     })
+
+/*Also ObjectIDs behave like Strings, and hence when comparing two ObjectIDs, you should use the Id1.equals(id2) syntax.*/
+
     .put(authenticate.verifyUser, (req, res, next) => {
         Dishes.findById(req.params.dishId)
             .then((dish) => {
-                if (dish != null && dish.comments.id(req.params.commentId) != null) {
+                if (dish != null && dish.comments.id(req.params.commentId) != null
+                    && dish.comments.id(req.params.commentId).author.equals(req.user._id)) {
                     if (req.body.rating) {
                         dish.comments.id(req.params.commentId).rating = req.body.rating;
                     }
@@ -183,9 +202,14 @@ dishRouter.route('/:dishId/comments/:commentId')
                     }
                     dish.save()
                         .then((dish) => {
-                            res.statusCode = 200;
-                            res.setHeader('Content-Type', 'application/json');
-                            res.json(dish);
+                            Dishes.findById(dish._id)
+                                .populate('comments.author')
+                                .then((dish) => {
+                                    res.statusCode = 200;
+                                    res.setHeader('Content-Type', 'application/json');
+                                    res.json(dish);
+                                })
+
                         }, (err) => next(err));
                 }
                 else if (dish == null) {
@@ -193,24 +217,38 @@ dishRouter.route('/:dishId/comments/:commentId')
                     err.status = 404;
                     return next(err);
                 }
-                else {
+                else if (dish.comments.id(req.params.commentId) == null){
                     err = new Error('Comment ' + req.params.commentId + ' not found');
                     err.status = 404;
                     return next(err);
                 }
+                else {
+                    err = new Error('you are not authorized to update this comment');
+                    err.status = 403;
+                    return next(err);  
+
+                }
+
             }, (err) => next(err))
             .catch((err) => next(err));
     })
     .delete(authenticate.verifyUser, (req, res, next) => {
+     
         Dishes.findById(req.params.dishId)
             .then((dish) => {
-                if (dish != null && dish.comments.id(req.params.commentId) != null) {
+                if (dish != null && dish.comments.id(req.params.commentId) != null
+                    && dish.comments.id(req.params.commentId).author.equals(req.user._id))
+                {
                     dish.comments.id(req.params.commentId).remove();
                     dish.save()
                         .then((dish) => {
-                            res.statusCode = 200;
-                            res.setHeader('Content-Type', 'application/json');
-                            res.json(dish);
+                            Dishes.findById(dish._id)
+                            .populate('comments.author')
+                            .then((dish) => {
+                                res.statusCode = 200;
+                                res.setHeader('Content-Type', 'application/json');
+                                res.json(dish);
+                            })
                         }, (err) => next(err));
                 }
                 else if (dish == null) {
@@ -218,10 +256,16 @@ dishRouter.route('/:dishId/comments/:commentId')
                     err.status = 404;
                     return next(err);
                 }
-                else {
+                else if (dish.comments.id(req.params.commentId) == null) {
                     err = new Error('Comment ' + req.params.commentId + ' not found');
                     err.status = 404;
                     return next(err);
+                }
+                else {
+                    err = new Error('you are not authorized to delete this comment');
+                    err.status = 403;
+                    return next(err);
+
                 }
             }, (err) => next(err))
             .catch((err) => next(err));
